requirements.txt
openai>=1.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tenacity>=8.2.0
lxml>=4.9.0
PyYAML>=6.0.1
streamlit>=1.30.0